server_aws/src/requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
pydantic==1.10.9
langchain==0.0.242
numpy==1.23.4
openai==0.27.8
faiss-cpu==1.7.4
sentence-transformers==2.2.2
PyMuPDF==1.22.3
python-dotenv==1.0.0
supabase==1.0.3
tqdm==4.64.1
python-multipart==0.0.6
PyJWT==2.8.0
